added index.html template for product, started show.html template for product, updated products controller with routes to index and show for products.

+++ db/inn_ventory.sql
@@ -15,8 +15,8 @@ CREATE TABLE products (
     name VARCHAR(255),
     category VARCHAR(255),
     in_stock INT,
-    cost_price INT,
-    sale_price INT,
+    cost_price FLOAT,
+    sale_price FLOAT,
     description TEXT,
     supplier_id INT REFERENCES suppliers(id)
 );
